requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pinecone>=5.0.0
openai>=1.6.0
anthropic>=0.18.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pytest>=7.0.0
